write_witness_state and read_actions subroutines
Implementation of the main routines that will serve to read actions and write state information into the database.
This will be done parallely for each rank, so the arrays in the database will be prefixed with mpirank_.
Arrays read from the database are then deleted, so that the poll_tensor function waits on the Python model to write the actions at the next time step.