Implemented the option to conver a inst file to a restat file
Heavy reestructuration of the class structure, splitting some stuff
in more subroutines and better encapsulation.

Started to implement the option to use a mesh already generated by the user
and not to use the mesh generated by the tool for viz.
Nonetheless still lot of work required, since is complex.